[TAN-1480] Separate modal code while still sharing much code
